requirements.txt
anthropic
openai
matplotlib
google-generativeai
pandas
numpy
python-dotenv
Pillow
pillow-heif
jupyter
ipywidgets
